Lets get this thing back to work

+++ src/database/Auction.java
@@ -7,12 +7,12 @@ public class Auction {
 
     private final int sellerID;
     private final int itemID;
-    private final int reserve;
+    private final float reserve;
     private final Date startTime;
     private final Date endTime;
     private ArrayList<Bid> bidList = new ArrayList<Bid>();
 
-    public Auction(int sellerID, int itemID, int reserve, Date startTime, Date endTime) {
+    public Auction(int sellerID, int itemID, float reserve, Date startTime, Date endTime) {
         this.sellerID = sellerID;
         this.itemID = itemID;
         this.reserve = reserve;
